Format code, move optimized array functions to jl-sys to keep jlrs unaware of these implementation details
