+++ jl_sys/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod bindings;
 pub mod gc_frame;
+pub mod inlined;
 pub mod types;
 
 pub use bindings::*;
